+++ tato.py
@@ -1,0 +1,4 @@
+tato = {1, 2, 4, 3}
+
+for tat in tato:
+    print(tat)